增加业务员审批，技术员打印装箱清单

+++ OpenAuth.Core/OpenAuth.App/Material/Request/AddOrUpdateQuotationReq.cs
@@ -94,7 +94,7 @@ namespace OpenAuth.App.Material.Request
         /// <summary>
         ///报价单审批状态
         /// </summary>
-        public int? QuotationStatus { get; set; }
+        public decimal? QuotationStatus { get; set; }
 
         /// <summary>
         ///备注
